Refactor to use only react context + render props (#1)
Gets rid of the singleton store and observable dependency in favor of React context + nesting

The most basic way to use this is just using a render prop that accepts state + setState and is almost the same as State from react-powerplug, but the whitespace limitations make it very difficult to work with. Providing the values as context seems cleaner for more complex documents even if it requires more local nesting. See demo/counter-child-function.mdx.